tms9995: use 16bit forms for fast8

+++ Library/include/tms9995/stdint.h
@@ -13,8 +13,8 @@ typedef signed char int8_t;
 typedef int16_t intptr_t;
 typedef uint16_t uintptr_t;
 
-typedef uint8_t uint_fast8_t;
-typedef int8_t int_fast8_t;
+typedef uint16_t uint_fast8_t;
+typedef int16_t int_fast8_t;
 typedef uint16_t uint_fast16_t;
 typedef int16_t int_fast16_t;
 typedef uint32_t uint_fast32_t;
